Split balanceFlows into its own rule; drop owned/initialBalance
- New rules/balance-flows.ts with `balanceFlows` (renamed from
  addressCoinFlows). Uses natural signs: negative = deduction,
  positive = inflow.
- TrackedCoin is now { coinType, balance, ownerAddress }. Primary vs
  derived distinction moves to the call site (trackPrimary vs
  trackDerived) instead of a flag on the class.
- coin-flows.ts shrinks to a backwards-compat wrapper: coinFlows and
  sponsorFlows are derived from balanceFlows by selecting an address
  and sign-flipping the deltas so outflow amounts come out positive.
  Every non-zero entry is included; callers that only care about
  spending filter by `amount > 0n` (the auto-approvals manager already
  does).
- Trimmed AI-style commentary on internal functions.
