+++ packages/wallet-sdk/src/transaction-analyzer/rules/balance-flows.ts
@@ -1,0 +1,386 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import { normalizeSuiAddress, normalizeStructTag } from '@mysten/sui/utils';
+import type { TransactionAnalysisIssue } from '../analyzer.js';
+import { createAnalyzer } from '../analyzer.js';
+import { bcs } from '@mysten/sui/bcs';
+import { commands } from './commands.js';
+import type { AnalyzedCommand, AnalyzedCommandArgument } from './commands.js';
+import { data } from './core.js';
+import { inputs } from './inputs.js';
+import { coinReservations, coins, gasCoins } from './coins.js';
+
+export interface CoinFlow {
+	coinType: string;
+	amount: bigint;
+}
+
+/**
+ * Per-address signed balance deltas. Negative = value left the address on
+ * net; positive = value arrived on net. Zero deltas are filtered out. For
+ * each coin type the sum across all addresses is zero, modulo coins consumed
+ * to an unknown destination.
+ */
+export interface BalanceFlowsResult {
+	byAddress: Record<string, CoinFlow[]>;
+}
+
+const SUI_FRAMEWORK = normalizeSuiAddress('0x2');
+
+export const balanceFlows = createAnalyzer({
+	dependencies: { data, commands, inputs, coins, gasCoins, coinReservations },
+	analyze:
+		() =>
+		async ({ data, commands, inputs, coins, gasCoins, coinReservations }) => {
+			const issues: TransactionAnalysisIssue[] = [];
+			const trackedCoins = new Map<string, TrackedCoin>();
+			const deltas = new Map<string, Map<string, bigint>>();
+
+			const sender = data.sender ? normalizeSuiAddress(data.sender) : null;
+			const gasOwner = data.gasData.owner
+				? normalizeSuiAddress(data.gasData.owner)
+				: (sender ?? null);
+
+			const normalizeAddress = (address: string | null): string | null =>
+				address == null ? null : normalizeSuiAddress(address);
+
+			const adjustDelta = (address: string | null, coinType: string, amount: bigint) => {
+				if (!address || amount === 0n) return;
+				let byCoin = deltas.get(address);
+				if (!byCoin) {
+					byCoin = new Map();
+					deltas.set(address, byCoin);
+				}
+				byCoin.set(coinType, (byCoin.get(coinType) ?? 0n) + amount);
+			};
+
+			// Primary: a coin entering the PTB from outside (input Object, gas
+			// coin, reservation, redeem_funds result). Debits the owner up
+			// front; the credit comes back via transfer-to-self or implicit
+			// return at end-of-PTB.
+			const trackPrimary = (key: string, coin: TrackedCoin) => {
+				trackedCoins.set(key, coin);
+				adjustDelta(coin.ownerAddress, coin.coinType, -coin.balance);
+			};
+
+			// Derived: a coin produced by redistributing existing tracked value
+			// (split, conversion, withdraw_all, coin::zero). No delta change.
+			const trackDerived = (key: string, coin: TrackedCoin) => {
+				trackedCoins.set(key, coin);
+			};
+
+			const trackedCoinKey = (ref: AnalyzedCommandArgument): string | null => {
+				switch (ref.$kind) {
+					case 'GasCoin':
+						return 'gas';
+					case 'Object':
+						return `input:${ref.index}`;
+					case 'Result':
+						return `result:${ref.index[0]},${ref.index[1]}`;
+					case 'Unknown':
+					case 'Pure':
+					case 'Withdrawal':
+						return null;
+				}
+			};
+
+			const getTrackedCoin = (ref: AnalyzedCommandArgument): TrackedCoin | null => {
+				const key = trackedCoinKey(ref);
+				return key ? (trackedCoins.get(key) ?? null) : null;
+			};
+
+			const splitCoin = (command: Extract<AnalyzedCommand, { $kind: 'SplitCoins' }>) => {
+				const coin = getTrackedCoin(command.coin);
+				if (!coin) return;
+
+				if (!command.amounts.every((a) => a.$kind === 'Pure')) {
+					coin.consume();
+					return;
+				}
+
+				const amounts = command.amounts.map((a) => {
+					if (a.$kind !== 'Pure') throw new Error('Expected pure value');
+					return BigInt(bcs.u64().fromBase64(a.bytes));
+				});
+
+				coin.balance -= amounts.reduce((a, b) => a + b, 0n);
+
+				amounts.forEach((amount, i) => {
+					trackDerived(
+						`result:${command.index},${i}`,
+						new TrackedCoin(coin.coinType, amount, coin.ownerAddress),
+					);
+				});
+			};
+
+			const mergeCoins = (command: Extract<AnalyzedCommand, { $kind: 'MergeCoins' }>) => {
+				const sources = command.sources.map(getTrackedCoin);
+				const amount = sources.reduce((a, c) => a + (c?.balance ?? 0n), 0n);
+				for (const src of sources) src?.consume();
+				const dest = getTrackedCoin(command.destination);
+				if (!dest) return;
+				dest.balance += amount;
+			};
+
+			const transferObjects = (command: Extract<AnalyzedCommand, { $kind: 'TransferObjects' }>) => {
+				const address =
+					command.address.$kind === 'Pure' ? bcs.Address.fromBase64(command.address.bytes) : null;
+				const dest = normalizeAddress(address);
+				for (const obj of command.objects) {
+					const tracked = getTrackedCoin(obj);
+					if (tracked) {
+						adjustDelta(dest, tracked.coinType, tracked.balance);
+						tracked.consume();
+					}
+				}
+			};
+
+			const getCoinTypeFromTypeArgs = (
+				command: Extract<AnalyzedCommand, { $kind: 'MoveCall' }>,
+			): string | null => {
+				const typeArg = command.command.typeArguments?.[0];
+				return typeArg ? normalizeStructTag(typeArg) : null;
+			};
+
+			const handleFrameworkMoveCall = (
+				command: Extract<AnalyzedCommand, { $kind: 'MoveCall' }>,
+			): boolean => {
+				const pkg = normalizeSuiAddress(command.command.package);
+				if (pkg !== SUI_FRAMEWORK) return false;
+
+				const mod = command.command.module;
+				const fn = command.command.function;
+				const coinType = getCoinTypeFromTypeArgs(command);
+
+				if (fn === 'redeem_funds' && (mod === 'coin' || mod === 'balance')) {
+					const arg = command.arguments[0];
+					if (arg?.$kind !== 'Withdrawal') return false;
+					const owner =
+						arg.withdrawFrom === 'Sender'
+							? (sender ?? null)
+							: (normalizeAddress(gasOwner) ?? sender);
+					trackPrimary(
+						`result:${command.index},0`,
+						new TrackedCoin(arg.coinType, arg.amount, normalizeAddress(owner)),
+					);
+					return true;
+				}
+
+				if (fn === 'send_funds' && (mod === 'coin' || mod === 'balance')) {
+					if (command.arguments.length < 2) return false;
+					const tracked = getTrackedCoin(command.arguments[0]);
+					const addrArg = command.arguments[1];
+					const destAddress =
+						addrArg.$kind === 'Pure' ? bcs.Address.fromBase64(addrArg.bytes) : null;
+					if (tracked) {
+						adjustDelta(normalizeAddress(destAddress), tracked.coinType, tracked.balance);
+						tracked.consume();
+					}
+					return true;
+				}
+
+				if (
+					(fn === 'into_balance' && mod === 'coin') ||
+					(fn === 'from_balance' && mod === 'coin')
+				) {
+					// Value-preserving re-key; no new deduction.
+					if (command.arguments.length < 1) return false;
+					const oldKey = trackedCoinKey(command.arguments[0]);
+					const tracked = oldKey ? trackedCoins.get(oldKey) : undefined;
+					if (tracked) {
+						trackedCoins.delete(oldKey!);
+						trackedCoins.set(`result:${command.index},0`, tracked);
+					}
+					return true;
+				}
+
+				if (
+					(fn === 'split' && (mod === 'coin' || mod === 'balance')) ||
+					(fn === 'take' && mod === 'coin')
+				) {
+					if (command.arguments.length < 2) return false;
+					const source = getTrackedCoin(command.arguments[0]);
+					if (!source) return true;
+
+					const amountArg = command.arguments[1];
+					if (amountArg.$kind !== 'Pure') {
+						source.consume();
+						return true;
+					}
+
+					const amount = BigInt(bcs.u64().fromBase64(amountArg.bytes));
+					source.balance -= amount;
+					trackDerived(
+						`result:${command.index},0`,
+						new TrackedCoin(source.coinType, amount, source.ownerAddress),
+					);
+					return true;
+				}
+
+				if (fn === 'withdraw_all' && mod === 'balance') {
+					if (command.arguments.length < 1) return false;
+					const source = getTrackedCoin(command.arguments[0]);
+					if (source) {
+						trackDerived(
+							`result:${command.index},0`,
+							new TrackedCoin(source.coinType, source.balance, source.ownerAddress),
+						);
+						source.balance = 0n;
+					}
+					return true;
+				}
+
+				if (
+					(fn === 'join' && (mod === 'coin' || mod === 'balance')) ||
+					(fn === 'put' && mod === 'coin')
+				) {
+					if (command.arguments.length < 2) return false;
+					const dest = getTrackedCoin(command.arguments[0]);
+					const source = getTrackedCoin(command.arguments[1]);
+					if (!source) return true;
+					if (dest) dest.balance += source.balance;
+					source.consume();
+					return true;
+				}
+
+				if (fn === 'zero' && (mod === 'coin' || mod === 'balance')) {
+					if (coinType) {
+						trackDerived(`result:${command.index},0`, new TrackedCoin(coinType, 0n, null));
+					}
+					return true;
+				}
+
+				// TODO: coin::mint / coin::mint_and_transfer (violates conservation),
+				// 0x2::pay module (split, split_and_transfer, join_vec).
+				return false;
+			};
+
+			const suiType = normalizeStructTag('0x2::sui::SUI');
+			const normalizedGasOwner = normalizeAddress(gasOwner);
+
+			const gasCoinsBalance = gasCoins.reduce((a, c) => a + c.balance, 0n);
+			const gasBalance = gasCoinsBalance + coinReservations.reduce((a, r) => a + r.balance, 0n);
+
+			// Empty gas payment: attribute the budget to the sender since the tx
+			// hasn't been resolved to a payer yet.
+			const gasAttributionOwner = gasBalance > 0n ? normalizedGasOwner : sender;
+
+			trackedCoins.set('gas', new TrackedCoin(suiType, gasBalance, gasAttributionOwner));
+
+			if (gasCoinsBalance > 0n) {
+				adjustDelta(normalizedGasOwner, suiType, -gasCoinsBalance);
+			}
+			for (const r of coinReservations) {
+				adjustDelta(normalizeAddress(r.owner), suiType, -r.balance);
+			}
+
+			if (data.gasData.budget) {
+				trackedCoins.get('gas')!.balance -= BigInt(data.gasData.budget);
+			} else {
+				issues.push({ message: 'Gas budget not set in Transaction' });
+			}
+
+			for (const input of inputs) {
+				if (input.$kind === 'Object' && coins[input.object.objectId]) {
+					const coin = coins[input.object.objectId];
+					trackPrimary(
+						`input:${input.index}`,
+						new TrackedCoin(coin.coinType, coin.balance, normalizeAddress(coin.ownerAddress)),
+					);
+				}
+			}
+
+			for (const command of commands) {
+				switch (command.$kind) {
+					case 'SplitCoins':
+						splitCoin(command);
+						break;
+					case 'MergeCoins':
+						mergeCoins(command);
+						break;
+					case 'TransferObjects':
+						transferObjects(command);
+						break;
+					case 'MakeMoveVec':
+						command.elements.forEach((el) => {
+							getTrackedCoin(el)?.consume();
+						});
+						break;
+					case 'MoveCall':
+						if (!handleFrameworkMoveCall(command)) {
+							command.arguments.forEach((arg) => {
+								if (arg.accessLevel === 'read') return;
+								getTrackedCoin(arg)?.consume();
+							});
+						}
+						break;
+					case 'Upgrade':
+					case 'Publish':
+						break;
+					default:
+						throw new Error(`Unsupported command type: ${command.$kind}`);
+				}
+			}
+
+			if (issues.length) return { issues };
+
+			// Implicit return: every still-alive tracked coin credits its owner
+			// with its remaining balance. For the gas coin — which may have been
+			// funded by multiple parties — split the leftover proportionally.
+			for (const [key, coin] of trackedCoins) {
+				if (coin.balance <= 0n) continue;
+				if (key === 'gas' && gasBalance > 0n) {
+					const contributors: { owner: string | null; amount: bigint }[] = [];
+					if (gasCoinsBalance > 0n) {
+						contributors.push({ owner: normalizedGasOwner, amount: gasCoinsBalance });
+					}
+					for (const r of coinReservations) {
+						if (r.balance > 0n) {
+							contributors.push({ owner: normalizeAddress(r.owner), amount: r.balance });
+						}
+					}
+					let distributed = 0n;
+					contributors.forEach((c, i) => {
+						const share =
+							i === contributors.length - 1
+								? coin.balance - distributed
+								: (c.amount * coin.balance) / gasBalance;
+						adjustDelta(c.owner, suiType, share);
+						distributed += share;
+					});
+				} else {
+					adjustDelta(coin.ownerAddress, coin.coinType, coin.balance);
+				}
+			}
+
+			const byAddress: Record<string, CoinFlow[]> = {};
+			for (const [address, byCoin] of deltas) {
+				const flows: CoinFlow[] = [];
+				for (const [coinType, delta] of byCoin) {
+					if (delta !== 0n) flows.push({ coinType, amount: delta });
+				}
+				if (flows.length > 0) byAddress[address] = flows;
+			}
+
+			return {
+				result: { byAddress } satisfies BalanceFlowsResult,
+			};
+		},
+});
+
+class TrackedCoin {
+	coinType: string;
+	balance: bigint;
+	ownerAddress: string | null;
+
+	constructor(coinType: string, balance: bigint, ownerAddress: string | null) {
+		this.coinType = coinType;
+		this.balance = balance;
+		this.ownerAddress = ownerAddress;
+	}
+
+	consume() {
+		this.balance = 0n;
+	}
+}
