+++ packages/wallet-sdk/src/transaction-analyzer/rules/index.ts
@@ -2,7 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { accessLevel } from './accessLevel.js';
-import { addressCoinFlows, coinFlows, sponsorFlows } from './coin-flows.js';
+import { balanceFlows } from './balance-flows.js';
+import { coinFlows, sponsorFlows } from './coin-flows.js';
 import { coinValues } from './coin-value.js';
 import { coinReservations, coins, gasCoins } from './coins.js';
 import { commands } from './commands.js';
@@ -13,8 +14,8 @@ import { objectIds, objects, objectsById, ownedObjects } from './objects.js';
 
 export const analyzers = {
 	accessLevel,
-	addressCoinFlows,
 	balanceChanges,
+	balanceFlows,
 	bytes,
 	coinFlows,
 	coinReservations,
